requirements.txt
# Core Framework
google-adk>=1.0.0
python-a2a[all]>=0.1.0

# Web Framework
flask>=2.3.0
fastapi>=0.104.0
uvicorn>=0.24.0
websockets>=12.0

# Database & Storage
sqlalchemy>=2.0.0
redis>=5.0.0
weaviate-client>=3.25.0
neo4j>=5.15.0
psycopg2-binary>=2.9.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
pydantic>=2.5.0
sentence-transformers>=2.2.0

# AI/ML
openai>=1.0.0
langchain>=0.1.0
anthropic>=0.7.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
aiohttp>=3.9.0
asyncio-mqtt>=0.13.0
networkx>=3.2.0

# Development Tools
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0

# Monitoring & Logging
prometheus-client>=0.19.0
structlog>=23.2.0
